feat: remove DOI component as it doesn't add much functionality
