+++ src/components/Doi/Doi.tsx
@@ -8,6 +8,9 @@ import CitationsChart from '../CitationsChart/CitationsChart'
 import DoiMetadata from '../DoiMetadata/DoiMetadata'
 import { compactNumbers } from '../../utils/helpers'
 
+
+import DoiRelatedContent from '../DoiRelatedContent/DoiRelatedContent'
+
 import UsageChart from '../UsageChart/UsageChart'
 
 type Props = {
@@ -20,102 +23,16 @@ const DoiPresentation: React.FunctionComponent<Props> = ({doi}) => {
       No content found.
     </Alert>
   )
-  
-  const formattedCitation = () => { 
-    const [selectedOption, setSelectedOption] = React.useState('')
 
-    return (
-      <div>
-        <div id="citation" className="input-group pull-right">
-          <select className="cite-as" onChange={e => setSelectedOption(e.target.value)} >
-              <option value="apa">APA</option>
-              <option value="harvard-cite-them-right">Harvard</option>
-              <option value="modern-language-association">MLA</option>
-              <option value="vancouver">Vancouver</option>
-              <option value="chicago-fullnote-bibliography">Chicago</option>
-              <option value="ieee">IEEE</option>
-          </select>
-        </div>
-        <CitationFormatter id={doi.doi} input={doi.formattedCitation} locale="en" style={selectedOption}></CitationFormatter>
-      </div>
-    )
-  }
-
-  const style = {
-    fontWeight: 600,  
-    color:'#1abc9c',
-    fontSize: '25px',
-    padding: 0,
-    margin: '0 0 .35em 10px',
-  }
-
-  const citationsTabLabel = Pluralize({count: compactNumbers(doi.citationCount), singular:'Citation', style:style, showCount:true}) 
-  const viewsTabLabel = Pluralize({count: compactNumbers(doi.viewCount), singular:'View', style:style, showCount:true}) 
-  const downloadsTabLabel = Pluralize({count: compactNumbers(doi.downloadCount), singular:'Download', style:style, showCount:true}) 
- 
-  const analyticsBar = () => {
-    return (
-      <div className="panel panel-transparent">
-        <div className="panel-body tab-content nav-tabs-member">
-          <Tabs  id="over-time-tabs">
-            {doi.citationCount > 0 && 
-              <Tab className="citations-over-time-tab" eventKey="citationsOverTime" title={citationsTabLabel}>
-                <CitationsChart data={doi.citationsOverTime} publicationYear={doi.publicationYear} citationCount={doi.citationCount}></CitationsChart>
-              </Tab>
-            }
-            {doi.viewCount > 0 && 
-              <Tab className="views-over-time-tab" eventKey="viewsOverTime" title={viewsTabLabel}>
-                <UsageChart data={doi.viewsOverTime} counts={doi.viewCount} publicationYear={doi.publicationYear} type="View"/> 
-              </Tab>
-            }
-            {doi.downloadCount > 0 && 
-              <Tab className="downloads-over-time-tab" eventKey="downloadsOverTime" title={downloadsTabLabel}>
-                <UsageChart data={doi.downloadsOverTime} counts={doi.downloadCount} publicationYear={doi.publicationYear} type="Download" />
-              </Tab>
-            }
-          </Tabs>
-        </div>
-      </div>
-    )
-  }
-
-// eslint-disable-next-line no-unused-vars
-  const relatedContent = () => {
-    const referencesTabLabel = Pluralize({count: compactNumbers(doi.references.nodes.length), singular:'Reference', style:style,showCount:true}) 
-    const citationsTabLabel = Pluralize({count: compactNumbers(doi.citations.nodes.length), singular:'Citation', style:style,showCount:true}) 
-
-    return (
-      <div className="panel panel-transparent">
-        <div className="panel-body tab-content nav-tabs-member">
-          <Tabs id="related-content-tabs">
-            {doi.citations.nodes.length > 0 && 
-              <Tab className="citations-list" eventKey="citationsList" title={citationsTabLabel}>
-                {/* <RelatedContentList dataInput={doi.citations} /> */}
-                <p>This feature will be implemented later in 2020. <a href="https://portal.productboard.com/71qotggkmbccdwzokuudjcsb/c/35-common-doi-search" target="_blank" rel="noreferrer">Provide input</a></p>
-
-              </Tab>
-            }
-            {doi.references.nodes.length > 0 && 
-              <Tab className="references-list" eventKey="referencesList" title={referencesTabLabel}>
-                {/* <RelatedContentList dataInput={doi.references} /> */}
-                <p>This feature will be implemented later in 2020. <a href="https://portal.productboard.com/71qotggkmbccdwzokuudjcsb/c/35-common-doi-search" target="_blank" rel="noreferrer">Provide input</a></p>
-
-              </Tab>
-            }
-          </Tabs>
-        </div>
-      </div>
-    )
-  }
 
   return (
     <div key={doi.id} className="panel panel-transparent">
       <h2 className="member-results">{doi.doi}</h2>
       <DoiMetadata metadata={doi}></DoiMetadata>
       <br/>
-      {formattedCitation()}
+      {/* {formattedCitation()}
       {analyticsBar()}
-      {/* {relatedContent()} */}
+      {relatedContent()} */}
     </div>
   )
 }
